+++ packages/components-js/projects/jsdom-polyfill/tests/unit/specs/jsdom-build.spec.ts
@@ -4,7 +4,7 @@ import * as fs from 'fs';
 // TODO: irrelevant?
 it('should have no stylesheet injection warning', async () => {
   const componentsJsEntry = require.resolve('@porsche-design-system/components-js');
-  const filePath = path.resolve(componentsJsEntry, '../jsdom-polyfill/index.js');
+  const filePath = path.resolve(componentsJsEntry, '../../jsdom-polyfill/index.cjs');
   const jsdomBuild = fs.readFileSync(filePath, 'utf8');
 
   expect(