+++ packages/components-js/tests/e2e/puppeteer/specs/examples.e2e.ts
@@ -18,7 +18,7 @@ beforeEach(async () => {
 });
 afterEach(async () => await page.close());
 
-const filePath = path.resolve(require.resolve('@porsche-design-system/components-js'), '../../../public/index.html');
+const filePath = path.resolve(require.resolve('@porsche-design-system/components-js'), '../../../../public/index.html');
 const fileContent = fs.readFileSync(filePath, 'utf8');
 
 const [, rawOptions] = /<select onchange.*([\s\S]*?)<\/select>/.exec(fileContent) || [];