+++ packages/components-js/tests/vrt/puppeteer/specs/modal.vrt.ts
@@ -9,7 +9,7 @@ import {
   forceFocusState,
   forceHoverState,
   getBodyMarkup,
-  GetMarkup,
+  type GetMarkup,
   setContentWithDesignSystem,
 } from '../helpers';
 
@@ -18,6 +18,16 @@ it.each(furtherExtendedViewports)('should have no visual regression for viewport
     await vrtTest(getVisualRegressionTester(viewport), 'modal', '/#modal', {
       scenario: async (page) => {
         await page.mouse.click(0, 0); // click top left corner of the page to remove focus on modal
+
+        // scroll modal once down and up to apply box-shadow
+        await page.evaluate(() => {
+          // page is initially 1px high and then resized which affects scroll based behavior
+          window.addEventListener('resize', () => {
+            const modal = document.querySelector('.scroll');
+            modal.scrollBy({ top: 5 });
+            modal.scrollBy({ top: -5 });
+          });
+        });
       },
     })
   ).toBeFalsy();