+++ packages/utilities/projects/styles/rollup.config.js
@@ -1,29 +1,31 @@
 import typescript from '@rollup/plugin-typescript';
 
 const input = 'src/index.ts';
-const outputDir = 'dist/js';
+const outputDir = 'dist';
 
 export default [
   {
     input,
     output: {
-      dir: outputDir,
+      dir: `${outputDir}/cjs`,
       format: 'cjs',
+      entryFileNames: '[name].cjs',
     },
-    plugins: [
-      typescript({
-        declaration: true,
-        declarationDir: `${outputDir}/types`,
-        rootDir: 'src',
-      }),
-    ],
+    plugins: [typescript()],
   },
   {
     input,
     output: {
       dir: `${outputDir}/esm`,
       format: 'esm',
+      entryFileNames: '[name].mjs',
     },
-    plugins: [typescript()],
+    plugins: [
+      typescript({
+        declaration: true,
+        declarationDir: `${outputDir}/esm`,
+        rootDir: 'src',
+      }),
+    ],
   },
 ];